resource: Refactor color class code into its own files
